Made a little bit of progress.  It was a bit slow today as I am still getting over this week and I'm at work today.
I had to rename Bootstrap.php to Bootstrap123.php.  When I tried to make this commit the first time, I was told this filename had been changed by changing the case of some letters and Netbeans was unable to handle that situation properly.  So, I just gave it a different name and then will rename it back to what it is supposed to be.