The current method of restricting access to each of the page requires the same code in almost every method of every models class.  It currently does not even factor in the different between users and admins.  I need to come up with a better solution for handling this situation.
Also, I made a note in the employee.php file about possibly using a MySQL 
stored procedure to accomplish a task which, as far as I see right now, would
require two PHP queries.

I'm not sure which would be the better solution as of yet.  I'll think on 
it tonight while I'm taking the Storm Spotter's class.