Created a few placeholder view files.
Created the form to display the pto days for the current user.

had to scrap my original employees index file which used tabs for different
employee functions.  I am disappointed about that because I thought it
was going to look really nice, but currently, I do no know how to implement
it under the current framework.