I am currently trying to work through the logic to calculate the times entered by the user.  This is one of the main functions of this project so I really need to get this functioning.
I have some basic ideas laid out in the comments of the files.  I am trying
to push through each baby step to try to get this to work.

My first step of this process is to build the methods to pull the data from
from the database for one employee.

The next step is to pull each line from the database and process it to the
appropriate variables.