tl:dr fixed an error with night runs and 24 hour shifts.  the night runs were not being properly handled and stored.
fixed the issue of night runs >= 3 hours not being entered as 8 hours.
Previously, there was no provision for this situation.

I was originally working on the administrators current time sheet view for
the entire service.  I forgot to stash or shelve my changes before starting this fix.

This fix entails a missed checked when determining how to calculate the
starting and ending times.  I had forgotten to account for when the shift
is a 24 hour shift AND a night run.  I fixed that.  

Before pushing this to remote, I realized that the system does not handle
the >3 hour night run situation.

Working on that now.

It appears that I have it all fixed for now, until someone finds another
way to break the system.