Added a banner for abuseipdb.com to views/home/index.php
It doesn't have to be on this page, but at least for now, I want it some
where on the website.  When this project goes to production, this banner
will probably have to be removed.

Also, if you have a little bit smaller, less intrusive banner idea, please
present it.