Working on the night time hours calculations.  Although I am making progress, I am still having issues.  I'm getting a couple of error messages in some places while getting incorrect values stored in the database in other areas. Not sure if it is because of bad calculations or bad values being passed.
Will have to troubleshoot this more when I get a chance.

I also was playing with testing frameworks.  I added "Karma" to this to
see if it would help any.  If it doesn't we can always take it back out.

I really need to start getting more sleep at night.