While looking through the code, I realized that I was not saving the "reason" entry.  I added that into the MySQL statement.  It'll be interesting to see how many errors that causes now.
I also made some progress on the recalculate times.  I now have it to the
point where it is trying to add the information to the database.

Currently, I have an issue with the way the times are handled.  I'll have to
investigate what I am doing wrong, but for now, it is time for bed.

I have to work tomorrow.