I have been working on getting the "update" button to work on the current pay period page.
In the processes, I realized the conflict and the increased complexity
caused by the time entry page using 1's and 0's to denote 'Y' and 'N'.

Since the database was already designed with 'Y' and 'N', I thought it
easier to change the javascript and the php to match the database.

Preliminary testing shows that it is working after the change.

I am now to the point where the recalculate routine is trying to write the
new information to the database.  This raised an issue that I had not
considered before now.  The Primary Key for the Payroll table is the
"Employee_Number" and the "Date_Time_In".

If I keep to my paradigm of "soft deleting" the old data and entering it
anew, then I have to rethink my Primary Key.

In this case, however, it might be ok to just update the information in the
database.  If I do this, however, I will need to add a new method instead
of reusing the old.