I think I finally got the recalculate method working properly.
Instead of doing the whole soft delete and re-entry, in this case, I simply
did an update query.  I didn't see any reason to clog up the database if
someone spams the "recalculate" button.

In fact, if I have all of my other logic working properly, this button should
not even ever be needed and may be removed later.